+++ app/Core/Session.php
@@ -4,11 +4,11 @@ namespace App\Core;
 
 class Session
 {
-    public static function set(string $key, string $value)
+    public static function set(string $key, $value)
     {
         $_SESSION[$key] = $value;
     }
-    public static function get(string $key): string | null
+    public static function get(string $key)
     {
         return $_SESSION[$key] ?? null;
     }