+++ app/Core/ValidationTrait.php
@@ -65,6 +65,16 @@ trait ValidationTrait
 
         return preg_match('/[!@#$%^&*()\-_=+{};:,<.>]/', $value);
     }
+    public function unique(string $value, string $table, string $column)
+    {
+        $db = App::resolve(Database::class);
+        $binding = [];
+        $binding[$column] = $value;
+        $record = $db->query("select {$column} from {$table} where {$column} = :{$column}", $binding)->find();
+
+        return $record == false;
+    }
+
 
     private $rules = [
         'string' => ':attribute must be a string',
@@ -80,6 +90,7 @@ trait ValidationTrait
         'digit' => ':attribute must have at least 1 digit',
         'symbol' => ':attribute must have at least 1 symbol',
         'upper' => ':attribute must have at least 1 uppercase',
+        'unique' => 'This :attribute already exists. Try to login',
     ];