+++ app/Models/User.php
@@ -2,6 +2,24 @@
 
 namespace App\Models;
 
-class User {
-    
+class User
+{
+    public string $email;
+    public string $name;
+    public string $username;
+    public string $password;
+
+    public function toArray()
+    {
+        return [
+            'email' => $this->email,
+            'name' => $this->name,
+            'username' => $this->username,
+            'password' => $this->password,
+        ];
+    }
+
+    public $casts = [
+        'password'
+    ];
 }