+++ app/controllers/AuthController.php
@@ -2,7 +2,10 @@
 
 namespace App\Controllers;
 
+use App\Core\Auth;
+use App\Core\Session;
 use App\Core\Validator;
+use App\Models\User;
 use App\Repositories\AuthRepository;
 use Exception;
 
@@ -29,40 +32,18 @@ class AuthController extends Controller
     // Register New User
     public function store()
     {
-        $data = $_POST;
-
-        $errors = [];
-
-        $validator = new Validator([
-            'name' => ['string', 'max:100', 'min:3'],
-            'username' => ['string', 'max:100', 'min:3'],
-            'email' => ['string', 'max:100', 'min:4', 'email'],
-            'password' => ['string', 'min:8', 'upper', 'symbol', 'digit'],
-        ]);
-
-        $validator->validate($data);
-        if (!$validator->isValid()) {
-            $errors = $validator->getMessages();
-        } else {
-            $data = $validator->validated();
-            $data['password'] = password_hash($data["password"], PASSWORD_DEFAULT);
-            AuthRepository::store($data);
-            $data = [];
-            $errors = [];
-            unset($_POST);
-            header("Location: /posts");
-        }
-
-        view("auth.register", [
-            'data' => $data,
-            'errors' => $errors,
-            ...$this->data,
-        ]);
+        $user = new User();
+        $user->email = $_POST['email'];
+        $user->name = $_POST['name'];
+        $user->username = $_POST['username'];
+        $user->password = $_POST['password'];
+        Auth::register($user);
     }
 
     // Render Author Profile Page
     public function show()
     {
+        dd(Session::get('user'));
         view("auth.profile");
     }